Added a new container (data_pool) and removed vector
Data Pool supports the same methods as queue and stack, but the pop
order is unordered (the only guarantee is that any data pushed will also
be popped before the container fails popping).

ccl::vector was removed because it was too impractical, both in that it
was very cache unfriendly and there are better ways to implement a
concurrent vector.